pyproject.toml
[project]
name = "mesh-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "numpy>=2.4.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scipy>=1.17.0",
    "trimesh>=4.11.0",
    "tripo3d>=0.3.10",
    "pyglet<2",
]
